+++ packages/interfaces/CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- (MINOR): Modification of `invoke` (`IAppliance`) to consume a PayloadArray.
+- (MAJOR): Modification of `invoke` (`IAppliance`) to return a PayloadArray.
 
 ### Added
 - `setup` and `teardown` abstract methods to IAppliance.